# src/main/resources/v1/eric-oss-auto-config-consistency-openapi.yaml
...
openapi: "3.0.2"
servers:
  - url: http://eacc/v1
info:
  x-api-id: 8608e72f-121d-4049-be33-7eb2d3cf5594
  x-audience: external-public
  description: |
    Ericsson Automated Configuration Consistency (EACC) provides self-service analytics 
    to enable users to uncover baseline consistency errors. EACC supports the scheduling 
    of CM consistency audits which detect inconsistencies based on user-supplied rules. 
    These inconsistencies can be corrected manually by the user. 
    Any changes can also be manually rolled back by the user.
  version: 0.1.5
  title: Ericsson Automated Configuration Consistency
  license:
    name: COPYRIGHT Ericsson 2024
    url: https://www.ericsson.com/en/legal
  termsOfService: https://www.ericsson.com/en/legal
  contact:
    email: pdlonteamq@pdl.internal.ericsson.com
...
paths:
...
  /v1/scopes:
    get:
      tags:
        - scopes
      summary: Get Metadata of all Scopes.
      description: Get metadata for all Scopes. An empty list is returned if no Scopes exist.
      operationId: listScopes
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ScopeMetadata'
              example:
                [
                  {
                    "id": "dedbfc30-41d5-46ff-ad39-ba36aebe2dbd",
                    "scopeName": "market1",
                    "uri": "v1/scopes/dedbfc30-41d5-46ff-ad39-ba36aebe2dbd"
                  },
                  {
                    "id": "6e57b373-cdb5-4d05-961f-17bef897ba9a",
                    "scopeName": "market2",
                    "uri": "v1/scopes/6e57b373-cdb5-4d05-961f-17bef897ba9a"
                  }
                ]
        default:
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
              example: {
                "title": "Failed to get scopes.",
                "status": 500,
                "detail": "Database operation failed."
              }
    post:
      tags:
        - scopes
      summary: Create a new Scope.
      description: Create a new Scope comprised of the Scope name and a CSV file containing the ManagedElement FDNs.
      operationId: createScope
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CreateScope'
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScopeMetadata'
              example:
                {
                  "id": "dedbfc30-41d5-46ff-ad39-ba36aebe2dbd",
                  "scopeName": "market1",
                  "uri": "v1/scopes/dedbfc30-41d5-46ff-ad39-ba36aebe2dbd"
                }
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
              example: {
                "title": "File does not match the csv format.",
                "status": 400,
                "detail": "Fill in all required fields with the appropriate information."
              }
        "413":
          $ref: '#/components/responses/413'
        default:
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
              example: {
                "title": "Failed to create scope.",
                "status": 500,
                "detail": "Database operation failed."
              }
  /v1/scopes/{id}:
    get:
      tags:
        - scopes
      summary: Get a Scope.
      description: Returns the ManagedElement FDNs for the Scope in CSV format for the specified ID.
      operationId: getScope
      parameters:
        - $ref: '#/components/parameters/acceptInHeader'
        - $ref: '#/components/parameters/scopeId'
      responses:
        "200":
          description: OK
          content:
            text/csv:
              schema:
                type: string
                format: binary
              example: |
                fdn
                "SubNetwork=Europe,SubNetwork=Ireland,MeContext=NR03gNodeBRadio00001,ManagedElement=NR03gNodeBRadio00001"
                "SubNetwork=Europe,SubNetwork=Ireland,MeContext=NR03gNodeBRadio00002,ManagedElement=NR03gNodeBRadio00002"
                "SubNetwork=Europe,SubNetwork=Ireland,MeContext=NR03gNodeBRadio00003,ManagedElement=NR03gNodeBRadio00003"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
              example: {
                "title": "Error parsing ID. Provided ID is invalid.",
                "status": 400,
                "detail": "Provide a valid ID."
              }
        "404":
          $ref: '#/components/responses/404_scope'
        default:
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
              example: {
                "title": "Failed to get scope.",
                "status": 500,
                "detail": "Database operation failed."
              }
    put:
      tags:
        - scopes
      summary: Update a Scope.
      description: |
        Update a Scope for the specified ID by providing a CSV file containing the ManagedElement FDNs. 
        This overwrites the existing ManagedElements in the Scope.
        A Scope cannot be updated if it is used in a Job that is involved in an ongoing execution.
      operationId: updateScope
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UpdateScope'
      parameters:
        - $ref: '#/components/parameters/acceptInHeader'
        - $ref: '#/components/parameters/contentTypeInHeader'
        - $ref: '#/components/parameters/scopeId'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScopeMetadata'
              example:
                {
                  "id": "dedbfc30-41d5-46ff-ad39-ba36aebe2dbd",
                  "scopeName": "market1",
                  "uri": "v1/scopes/dedbfc30-41d5-46ff-ad39-ba36aebe2dbd"
                }
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
              example: {
                "title": "File does not match the csv format.",
                "status": 400,
                "detail": "Fill in all required fields with the appropriate information."
              }
        "404":
          $ref: '#/components/responses/404_scope'
        "409":
          description: Conflict
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
              example: {
                "title": "Cannot update scope as it is used by an ongoing execution.",
                "status": 409,
                "detail": "Scope in use."
              }
        "413":
          $ref: '#/components/responses/413'
        default:
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
              example: {
                "title": "Failed to update scope.",
                "status": 500,
                "detail": "Database operation failed."
              }
    delete:
      tags:
        - scopes
      summary: Delete a Scope
      description: Delete a Scope for the specified ID. A Scope can only be deleted if is not used in a Job.
      operationId: deleteScope
      parameters:
        - $ref: '#/components/parameters/acceptInHeader'
        - $ref: '#/components/parameters/scopeId'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
              example: {
                "title": "Error parsing ID. Provided ID is invalid.",
                "status": 400,
                "detail": "Provide a valid ID."
              }
        "404":
          $ref: '#/components/responses/404_scope'
        "409":
          description: Conflict
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
              example: {
                "title": "Scope with provided ID is used in a Job configuration.",
                "status": 409,
                "detail": "Only Scopes not associated with a Job configuration can be deleted."
              }
        default:
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
              example: {
                "title": "Failed to delete scope.",
                "status": 500,
                "detail": "Database operation failed."
              }
...
components:
...
  parameters:
...
    scopeId:
      name: id
      in: path
      description: ID of the Scope
      required: true
      schema:
        type: string
...
    acceptInHeader:
      name: Accept
      in: header
      required: false
      schema:
        type: string
        example: application/json
        default: application/json
    contentTypeInHeader:
      name: Content-Type
      in: header
      required: false
      schema:
        type: string
        example: application/json
        default: application/json

  # Common error responses
  responses:
    "404_scope":
      description: Not Found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ErrorDetails'
          example: {
            "title": "Scope with provided ID does not exist.",
            "status": 404,
            "detail": "Provide a valid Scope ID."
          }
    "413":
      description: Payload Too Large
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ErrorDetails'
          example: {
            "title": "Maximum upload size exceeded.",
            "status": 413,
            "detail": "File is too large: the request was rejected because its size (58534393) exceeds the configured maximum (20971520)"
          }
...
  schemas:
...
    ScopeMetadata:
      type: object
      required: [ id, scopeName, uri ]
      properties:
        id:
          type: string
          description: The Scope ID.
        scopeName:
          type: string
          description: The Scope name.
        uri:
          type: string
          description: The URI of the Scope.
    CreateScope:
      type: object
      required: [ scopeName, fileName ]
      properties:
        scopeName:
          type: string
          description: Use lower case alphanumeric, underscores and dashes up to a maximum of 255 characters.
        filename:
          type: string
          format: binary
          description: |
            A CSV file containing the ManagedElement FDNs. See the User Guide for information on the format of this file.
    UpdateScope:
      type: object
      required: [ fileName ]
      properties:
        fileName:
          type: string
          format: binary
          description: |
            A CSV file containing the ManagedElement FDNs. See the User Guide for information on the format of this file.
...
    ErrorDetails:
      description: Provides additional information in an error response.
      type: object
      required: [ title, status, detail ]
      properties:
        title:
          $ref: '#/components/schemas/ErrorTitle'
        status:
          $ref: '#/components/schemas/ErrorStatus'
        detail:
          $ref: '#/components/schemas/ErrorDetail'
...
    ErrorTitle:
      type: string
      description: Short, human-readable summary of the problem.
    ErrorStatus:
      type: integer
      description: HTTP status code.
    ErrorDetail:
      description: |
        Optional human-readable explanation of the problem. if present, ought to focus on helping the client 
        correct the problem, rather than giving debugging information.